+++ src/tko/settings/logger.py
@@ -4,6 +4,7 @@ from tko.settings.log_action import LogAction
 from tko.settings.listener_daily import DailyListener
 from tko.settings.listener_task import TaskListener
 from tko.settings.listener_week import WeekListener
+from tko.settings.listener_cache import CacheListener
 
 class Logger:
     instance: None | Logger = None
@@ -20,11 +21,15 @@ class Logger:
         self.daily = DailyListener()
         self.tasks = TaskListener()
         self.week = WeekListener()
+        self.cache = CacheListener()
 
     def set_log_files(self, log_file: str) -> Logger:
         if self.history is not None:
             return self
-        self.history = HistoryFile(log_file, [self.daily.listener, self.tasks.listener, self.week.listener])
+        self.history = HistoryFile(log_file, [self.daily.listener, 
+                                              self.tasks.listener, 
+                                              self.week.listener, 
+                                              self.cache.listener])
         # resume = self.week.resume()
         # print("\n".join([f"{x.weed_day} {x.elapsed}" for x in resume.values()]))
         return self